Added files via upload
Basically just modified each class. I guess the main thing is the change in the Event class where each time a new instance is made, the instance goes in to an arraylist. Had to modify The rest of the classes to incorporate the change to the Event class. Also added JLabels (Good Idea Wesley) to the calendar panel and the textfields and button are now in the calendar panel also.